[DDW-701]: Update log file rotation
