+++ source/main/index.js
@@ -3,8 +3,7 @@ import os from 'os';
 import path from 'path';
 import { app, BrowserWindow, shell } from 'electron';
 import { client } from 'electron-connect';
-import { rotateOldLogFiles } from '../common/utils/logging';
-import { Logger } from './utils/logging';
+import { Logger, rotateOldLogFiles } from './utils/logging';
 import { setupLogging, logSystemInfo } from './utils/setupLogging';
 import { getNumberOfEpochsConsolidated } from './utils/getNumberOfEpochsConsolidated';
 import { handleDiskSpace } from './utils/handleDiskSpace';
